@@ -19,65 +19,65 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^15.2.4",
-    "@angular/cdk": "^15.2.4",
-    "@angular/common": "^15.2.4",
-    "@angular/compiler": "^15.2.4",
-    "@angular/core": "^15.2.4",
-    "@angular/forms": "^15.2.7",
-    "@angular/material": "^15.2.7",
-    "@angular/platform-browser": "^15.2.4",
-    "@angular/platform-browser-dynamic": "^15.2.4",
-    "@angular/platform-server": "^15.2.7",
-    "@angular/router": "^15.2.7",
-    "@nguniversal/express-engine": "^15.2.1",
-    "bootstrap": "^5.2.3",
+    "@angular/animations": "^16.2.0",
+    "@angular/cdk": "^16.2.0",
+    "@angular/common": "^16.2.0",
+    "@angular/compiler": "^16.2.0",
+    "@angular/core": "^16.2.0",
+    "@angular/forms": "^16.2.0",
+    "@angular/material": "^16.2.0",
+    "@angular/platform-browser": "^16.2.0",
+    "@angular/platform-browser-dynamic": "^16.2.0",
+    "@angular/platform-server": "^16.2.0",
+    "@angular/router": "^16.2.0",
+    "@nguniversal/express-engine": "^16.2.0",
+    "bootstrap": "^5.3.1",
     "express": "^4.18.2",
-    "marked": "^4.3.0",
-    "ngx-markdown": "^15.1.2",
+    "marked": "^7.0.2",
+    "ngx-markdown": "^16.0.0",
     "prism": "^4.1.2",
     "prismjs": "^1.29.0",
-    "rxjs": "^7.8.0",
-    "tslib": "^2.5.0",
-    "zone.js": "^0.13.0"
+    "rxjs": "^7.8.1",
+    "tslib": "^2.6.1",
+    "zone.js": "^0.13.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^15.2.6",
-    "@angular-eslint/builder": "15.2.1",
-    "@angular-eslint/eslint-plugin": "15.2.1",
-    "@angular-eslint/eslint-plugin-template": "15.2.1",
-    "@angular-eslint/schematics": "15.2.1",
-    "@angular-eslint/template-parser": "15.2.1",
-    "@angular/cli": "^15.2.6",
-    "@angular/compiler-cli": "^15.2.7",
-    "@nguniversal/builders": "^15.2.1",
-    "@nguniversal/express-engine": "^15.2.1",
-    "@testing-library/angular": "^13.4.0",
-    "@testing-library/dom": "^9.2.0",
+    "@angular-devkit/build-angular": "^16.2.0",
+    "@angular-eslint/builder": "16.1.0",
+    "@angular-eslint/eslint-plugin": "16.1.0",
+    "@angular-eslint/eslint-plugin-template": "16.1.0",
+    "@angular-eslint/schematics": "16.1.0",
+    "@angular-eslint/template-parser": "16.1.0",
+    "@angular/cli": "^16.2.0",
+    "@angular/compiler-cli": "^16.2.0",
+    "@nguniversal/builders": "^16.2.0",
+    "@nguniversal/express-engine": "^16.2.0",
+    "@testing-library/angular": "^14.2.0",
+    "@testing-library/dom": "^9.3.1",
     "@testing-library/jasmine-dom": "^1.3.3",
     "@testing-library/user-event": "^14.4.3",
     "@types/express": "^4.17.17",
-    "@types/jasmine": "^4.3.1",
-    "@types/marked": "^4.0.8",
-    "@types/node": "^18.15.11",
-    "@typescript-eslint/eslint-plugin": "5.48.2",
-    "@typescript-eslint/parser": "5.48.2",
-    "eslint": "^8.33.0",
+    "@types/jasmine": "^4.3.5",
+    "@types/marked": "^5.0.1",
+    "@types/node": "^20.4.10",
+    "@typescript-eslint/eslint-plugin": "6.3.0",
+    "@typescript-eslint/parser": "6.3.0",
+    "eslint": "^8.47.0",
     "eslint-config-google": "^0.14.0",
-    "eslint-config-prettier": "^8.8.0",
+    "eslint-config-prettier": "^9.0.0",
     "eslint-plugin-json": "^3.1.0",
-    "eslint-plugin-jsonc": "^2.7.0",
-    "eslint-plugin-prettier": "^4.2.1",
+    "eslint-plugin-jsonc": "^2.9.0",
+    "eslint-plugin-prettier": "^5.0.0",
     "eslint-plugin-rxjs": "^5.0.3",
-    "jasmine-core": "^4.6.0",
-    "karma": "^6.4.1",
-    "karma-chrome-launcher": "^3.1.1",
-    "karma-coverage": "^2.2.0",
+    "jasmine-core": "^5.1.0",
+    "karma": "^6.4.2",
+    "karma-chrome-launcher": "^3.2.0",
+    "karma-coverage": "^2.2.1",
     "karma-jasmine": "^5.1.0",
-    "karma-jasmine-html-reporter": "^2.0.0",
-    "prettier": "^2.8.7",
+    "karma-jasmine-html-reporter": "^2.1.0",
+    "prettier": "^3.0.1",
     "prettier-eslint": "^15.0.1",
     "prettier-eslint-cli": "^7.1.0",
-    "typescript": "^4.9.5"
+    "typescript": "^5.1.6"
   }
 }
