@@ -5,15 +5,23 @@ import { MatMenuModule } from '@angular/material/menu';
 import { MatToolbarModule } from '@angular/material/toolbar';
 import { MenuComponent } from './components/menu/menu.component';
 import { MatButtonModule } from '@angular/material/button';
+import { MatSidenavModule } from '@angular/material/sidenav';
+import { MatListModule } from '@angular/material/list';
+import { RouterModule } from '@angular/router';
+import {LayoutModule} from '@angular/cdk/layout';
 
 @NgModule({
   declarations: [MenuComponent],
   imports: [
     CommonModule,
+    LayoutModule,
+    MatButtonModule,
     MatIconModule,
     MatMenuModule,
     MatToolbarModule,
-    MatButtonModule,
+    MatSidenavModule,
+    MatListModule,
+    RouterModule,
   ],
   exports: [MenuComponent],
 })
